[Refactor] 분리 가능한 useEffect 동작 hook으로 분리.
