+++ src/components/customlib/SetQueryString.js
@@ -2,6 +2,7 @@ import config from '../../config/config.json';
 
 export default function SetQueryString(endPoint, condition) {
   let targetUrl = config.BASE_URL + endPoint;
+
   condition.forEach(item => {
     const { category, value } = item;
     const replaceValueByType =
